+++ VS2019/audio_analysis_lib.vcxproj
@@ -27,14 +27,12 @@
     <ClCompile Include="..\src\impl\world_d4c_ap.cpp" />
     <ClCompile Include="..\src\impl\world_harvest_f0.cpp" />
     <ClCompile Include="..\src\impl\world_synthesis.cpp" />
-    <ClCompile Include="..\src\impl\world_synthesis_impl.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\define_macro.hpp" />
     <ClInclude Include="..\src\fft_component.hpp" />
     <ClInclude Include="..\src\fft_exception.hpp" />
     <ClInclude Include="..\src\impl\input_capture_impl.hpp" />
-    <ClInclude Include="..\src\impl\world_synthesis_impl.hpp" />
     <ClInclude Include="..\src\input_capture.hpp" />
     <ClInclude Include="..\src\lib_audio_analysis.h" />
     <ClInclude Include="..\src\window_functions.hpp" />
