+++ VS2019/audio_analysis_lib.vcxproj.filters
@@ -39,15 +39,9 @@
     <ClCompile Include="..\src\impl\world_synthesis.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\impl\world_synthesis_impl.cpp">
-      <Filter>ソース ファイル</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\impl\input_capture_impl.hpp">
-      <Filter>ヘッダー ファイル</Filter>
-    </ClInclude>
-    <ClInclude Include="..\src\impl\world_synthesis_impl.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
     <ClInclude Include="..\src\define_macro.hpp">
